Update value type

+++ plugin/path_config_test.go
@@ -40,8 +40,8 @@ func TestConfig(t *testing.T) {
 		"service_account_email":      "",
 		"identity_token_audience":    "",
 		"identity_token_ttl":         int64(0),
-		"rotation_window":            0,
-		"rotation_period":            0,
+		"rotation_window":            float64(0),
+		"rotation_period":            float64(0),
 		"rotation_schedule":          "",
 		"disable_automated_rotation": false,
 	}
